Add files via upload
in this project ,I worked on medical rep. sales dataset of features like region,product name,rep. ids,month,.... 
-added column for net gain , ROI(return on investment in %) to evaluate the performance of the representatives
-analyzed the data using excel pivot tables and created an interactive dashboard containing charts with synced slicers to select desired categories(region,rep,,time,...)
i have the uploaded the dataset, pivot tables and dashboard in 1 excel file(each in a seperate sheet)
feel free to download the excel file and see a screenshot of the dashboard in the uploads!